m10/ponderada 3/microsservices/service1/requirements.txt
fastapi
uvicorn
pydantic
pillow
rembg